Prevent potential stack alignment problems
The stack pointer should ideally always point at the last item in the stack.
If there are no items on the stack, the stackpointer should therefor point "outside" of memory.
This isn't possible without changing the init scripts though, as the stack pointer is used as-is.

But, more importantly here the stack should at least be aligned to eight octets.

see https://stackoverflow.com/a/28761856